+++ src/core/Face.cpp
@@ -54,7 +54,7 @@ std::string toStr(const Face& face) {
   }
 }
 
-std::pair<int, Face> parseFace(const std::string& str) {
+std::pair<size_t, Face> parseFace(const std::string& str) {
   if (str.empty()) return {0, {}};
   char first_char = str[0];
   if (first_char == 'U')
